Changed argument type for delete and add

+++ Asset Management/Core/ICore.java
@@ -7,10 +7,10 @@ import Items.ItemType;
 
 public interface ICore {
 	boolean addItem(ItemProp ip, ItemType it);
-	boolean deleteItem(long id);
+	boolean deleteItem(Integer id);
 	boolean addDepartment(String dept);
 	boolean addLocation(String loc);
-	boolean borrowItem(long id);
+	boolean borrowItem(Integer id);
 	List<ItemProp> getItemList(ItemType it);
 	
 	List<String> getLocations();
